turn scans into modules
