+++ README.md
@@ -1,5 +1,7 @@
 # rekon
-Rekon is a DNS reconiassiance tool that helps to build a domain enumeration inventory with associated certificates, DNS names, and IP addresses. The tool also performs some optional rudimentery scans, to build up a list of missing secuirty headers, PII found in responses, and HTTP status codes. 
+`rekon`` is an attack surface managent tool that helps to collect basic pubically visible asset information though DNS enumeration and a variety of passive probes. This tool is meant to be able to build an asset inventory from a top-level domain (e.g. acme.com) using free and open source solutions. For an example of the output please see the [example_output](./example_output/) folder in this repository.
+ 
+`rekon` uses a combination of active and passive techinques to gather information. You should only run this tool against sites for which you have permission to run an active scan.
 
 `rekon` is tested against Pyton 3.11.5 on OS X, but is should run anywhere.
 
@@ -15,7 +17,7 @@ Rekon is a DNS reconiassiance tool that helps to build a domain enumeration inve
 
 1. Install Dependencies
 
-    `pip install -r requirements.txt`
+    `pip3 install -r requirements.txt`
 
 # Configuration
 
@@ -35,7 +37,6 @@ Rekon uses a configuration file (rekon-config.json) to specify the scanning para
     "run_ip_scan": true, -- Provide a list of associated IP addresses
     "http_status_scan": true, -- Return the HTTP status of the domain
     "sec_headers_scan": true, -- Look for specific missing security headers
-    "run_pii_scan": true, -- See what PII may be returned by the default request.
 }
 ```
 
@@ -49,7 +50,7 @@ _Running the script over VPN is recommended_.
 
 1. Run rekon
 
-    `python rekon.py`
+    `python3 rekon.py`
 
 # Viewing the results
 
@@ -59,23 +60,23 @@ You can open the CSV file with a spreadsheet application or text editor to view 
 
 # Other reconnisance references
 
-Here's some other references that may be helpful to see how other approach this kind of attack surface reconnisance. 
+Here are some other references that may be helpful to see how other approach this kind of attack surface reconnisance. 
 
 ## How to blogs
 
 * [How to: Recon and Content Discovery](https://www.hackerone.com/ethical-hacker/how-recon-and-content-discovery) - HackerOne
 
 
-## Useful free recon websites
+## Useful free browser-based reconnaissance tools
 
 * [crt.sh](https://crt.sh/) -  Certification search, providing DNS enumeration.
 * [DNS dumpster](https://dnsdumpster.com/) - Domain research tool with very robust data sets.
 * [Shodan.io](https://www.shodan.io/) - Global search engine for internet connected devices.
 
-## Commercial tools
+## Commercial ASM Tools
 
 * [Security Trails](https://securitytrails.com/) - Building DNS and IP realted inventories and monitoring.
-* 
+* [HackerOne Attack Surface Reviuew](https://www.hackerone.com/attack-surface-review)
 
 # Open source tools
 
@@ -84,5 +85,6 @@ Here's some other references that may be helpful to see how other approach this 
 * [wafw00f](https://github.com/EnableSecurity/wafw00f) - Firewall fingerprinting tool.
 * [assetfinder](https://github.com/tomnomnom/assetfinder)
 * [OWASP Amass](https://github.com/owasp-amass/amass) - Attack surface network mapping.
+* [Attack Surface Mapper](https://github.com/superhedgy/AttackSurfaceMapper) - Requires API keys for multiple integrations
 
 [Here's a good blog article](https://blog.yeswehack.com/yeswerhackers/subdomains-tools-review-full-detailed-comparison/) by [Siz2dez](https://twitter.com/Six2dez1) that provides a more comprehensive list of subdomain tools.
