--- example_output/scan_results_20230922.csv
@@ -1,10 +1,0 @@
-CRTSH ID,Logged At,Not Before,Not After,Common Name,Issuer Name,firewall,associated_ips,http_response,missing_sec_headers,pii_found,exposed_page
-24560643,2016-07-14T07:30:08.461,2016-07-14T00:00:00,2018-07-14T23:59:59,*.example.com,"C=US, O=""thawte, Inc."", CN=thawte SSL CA - G2",,,999,,,DNT
-34083306,2016-09-23T09:27:06.583,2010-09-02T16:36:49,2011-10-01T16:31:00,*.hosted.jivesoftware.com,"emailAddress=pki_admin@sungard.com, O=SunGard Availability Services, CN=SAS Public CA v1",,,999,,,DNT
-34001389,2016-09-23T03:15:34.309,2010-09-02T16:37:43,2011-10-01T16:32:00,*.uat3.hosted.jivesoftware.com,"emailAddress=pki_admin@sungard.com, O=SunGard Availability Services, CN=SAS Public CA v1",,,999,,,DNT
-8506962125,,2023-01-27T01:21:18,2033-01-24T01:21:18,example.com,Issuer Not Found,Edgecast (Verizon Digital Media) WAF.,93.184.216.34,200,"Content-Security-Policy,Strict-Transport-Security,X-Frame-Options,Referrer-Policy",DNT,No result found
-9813012734,2023-07-03T07:04:15.787,2023-07-03T00:00:00,2023-10-01T23:59:59,sample.com,"C=US, ST=TX, L=Houston, O=""cPanel, Inc."", CN=""cPanel, Inc. Certification Authority""",Wordfence (Defiant) WAF.,23.235.219.27,200,"Content-Security-Policy,Strict-Transport-Security,X-Frame-Options,Referrer-Policy",DNT,No result found
-24564717,2016-07-14T07:55:01.55,2016-07-14T00:00:00,2017-07-14T23:59:59,www.example.com,"C=US, O=""thawte, Inc."", CN=thawte SSL CA - G2",Edgecast (Verizon Digital Media) WAF.,93.184.216.34,200,"Content-Security-Policy,Strict-Transport-Security,X-Frame-Options,Referrer-Policy",DNT,No result found
-8396709327,2023-01-13T13:18:21.987,2023-01-13T00:00:00,2024-02-13T23:59:59,www.example.org,"C=US, O=DigiCert Inc, CN=DigiCert TLS RSA SHA256 2020 CA1",Edgecast (Verizon Digital Media) WAF.,93.184.216.34,200,"Content-Security-Policy,Strict-Transport-Security,X-Frame-Options,Referrer-Policy",DNT,No result found
-5379478497,2021-10-09T03:14:32.356,2021-10-09T02:14:32,2022-01-07T02:14:31,www.sample.com,"C=US, O=Let's Encrypt, CN=R3",Wordfence (Defiant) WAF.,23.235.219.27,200,"Content-Security-Policy,Strict-Transport-Security,X-Frame-Options,Referrer-Policy",DNT,No result found
-9014232891,2023-03-29T09:21:41.625,2023-03-29T00:00:00,2023-06-27T23:59:59,www14.sample.com,"C=AT, O=ZeroSSL, CN=ZeroSSL ECC Domain Secure Site CA",DNT,,999,"Error: HTTPSConnectionPool(host='www14.sample.com', port=443): Max retries exceeded with url: / (Caused by NameResolutionError(""<urllib3.connection.HTTPSConnection object at 0x136466dd0>: Failed to resolve 'www14.sample.com' ([Errno 8] nodename nor servname provided, or not known)""))",,"An exception occurred HTTPSConnectionPool(host='www14.sample.com', port=443): Max retries exceeded with url: / (Caused by NameResolutionError(""<urllib3.connection.HTTPSConnection object at 0x136477fd0>: Failed to resolve 'www14.sample.com' ([Errno 8] nodename nor servname provided, or not known)""))"
