+++ rekon-config.json
@@ -1,17 +1,18 @@
 {
     "root_urls": [
-       "example.com"
+       "example.com",
+       "hackthissite.org",
+       "hbh.sh"
+
     ],
     "max_dns_records": 50,
     "run_ip_scan": true,
-    "run_pii_scan": false,
     "http_status_scan": true,
     "sec_headers_scan": true,
     "check_exposed_pages": true,
-    "run_firewall_scan": false,
+    "run_firewall_scan": true,
     "run_robots_txt_scan": true,
     "run_port_scan": false,
-    
     "ports": [
         43,
         8080
