+++ rekon.py
@@ -1,6 +1,6 @@
 # MIT License
 #
-# Copyright (c) Tim Kelly
+# Copyright (c) 2023 Tim Kelly
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -23,13 +23,6 @@
 import requests
 from datetime import datetime
 import pandas as pd
-import dns.resolver
-import numpy as np
-import subprocess
-import re
-from requests.exceptions import SSLError, ConnectionError, ReadTimeout
-from urllib3.exceptions import NewConnectionError
-import time
 import json
 
 # scanning modules function imports
@@ -37,6 +30,7 @@ from scan_modules.robots_txt_scanner import check_robots_txt
 from scan_modules.ip_collect_scanner import check_ip_addresses
 from scan_modules.sec_headers_scan import check_security_headers
 from scan_modules.http_result_scan   import check_http_response
+from scan_modules.firewal_scan import detect_firewall
 
 # Regular expressions and their corresponding types for PII patterns
 pii_patterns = [
@@ -115,91 +109,9 @@ def start_scan(urls_to_scan):
         # Append the data frame to the list
         dfs.append(df)
 
-# Function to scan for PII in HTTP responses and update deduplicated_df
-def scan_and_update_pii_data(deduplicated_df):
-    pii_data = []
-
-    for index, row in deduplicated_df.iterrows():
-        common_name = row["Common Name"].strip()
-        http_response_code = int(check_http_response(common_name))  # Convert to integer
-        if common_name and http_response_code < 400:
-            pii_data += scan_for_pii_in_http_response(common_name)
-
-    pii_df = pd.DataFrame(pii_data) if pii_data else pd.DataFrame(["No results"], columns=["Message"])
-
-    # Add the "pii_found" column to the original deduplicated_df
-    deduplicated_df["pii_found"] = deduplicated_df["Common Name"].map(
-        lambda x: next((row["found_pii"] for _, row in pii_df.iterrows() if row["Common Name"] == x), None)
-    )
-
-def scan_for_pii_in_http_response(common_name):
-    common_name = common_name.strip()
-    http_response_code = check_http_response(common_name)
-    pii_data = []
-
-    if common_name and http_response_code < 400:
-        try:
-            response = requests.get(f"https://{common_name}")
-            response.raise_for_status()
-
-            source_code = response.text
-
-            pii_found = scan_for_pii(source_code)
-
-            if pii_found:
-                formatted_pii = ", ".join(
-                    [f"{pii_value} ({pii_type})" for pii_value, pii_type in pii_found]
-                )
-                pii_data.append({"Common Name": common_name, "found_pii": formatted_pii})
-            else:
-                # Set the default value for pii_value
-                pii_data.append({"Common Name": common_name, "found_pii": "DNT"})
-
-        except requests.exceptions.RequestException as e:
-            print(f"Error for {common_name}: {e}")
-
-    return pii_data
-
-# Define the function to scan for PII in text
-def scan_for_pii(text):
-    found_pii = []
-    for pattern, pii_type in pii_patterns:
-        matches = re.findall(pattern, text)
-        if matches:
-            for pii_value in matches:
-                found_pii.append((pii_value, pii_type))
-    return found_pii
 
 def is_valid_domain(domain):
     return not domain.startswith("*.") and "." in domain
-
-def clean_firewall_name(name):
-    return re.sub(r"\x1b\[\d+;\d+m", "", name)
-
-# Define the function for firewall detection
-def detect_firewall(domain):
-    def identify_firewall(domain):
-        command = f"wafw00f {domain}"
-        result = subprocess.check_output(command, shell=True, text=True)
-        return result.strip()
-
-    firewall_result = None  # Initialize with None by default
-
-    if is_valid_domain(domain):
-        if "*" not in domain:
-            firewall_result = "DNT"  # Default in case the test does not run
-            firewall_result_raw = identify_firewall(domain)
-            firewall_lines = firewall_result_raw.split('\n')
-
-            for line in firewall_lines:
-                if "No WAF detected" in line:
-                    firewall_result = "No WAF detected"
-                    break
-                elif "is behind" in line:
-                    firewall_result = clean_firewall_name(line.split("is behind", 1)[1].strip())
-                    break
-
-    return firewall_result
 
 # Function to write DataFrame to a CSV file
 def write_to_csv(dataframe, filename):
@@ -224,13 +136,16 @@ deduplicated_df = aggregate_df.sort_values(by=["Common Name", "Not After"], asce
 dns_names_list = deduplicated_df["Common Name"].tolist()
 
 if config["run_firewall_scan"]:
-    print(f"Starting firewall scan . . .")
 
+    print(f"Starting firewall scan . . .")
     firewall_results = []
 
     for dns_name in dns_names_list:
-        firewall_result = detect_firewall(dns_name)
-        firewall_results.append(firewall_result)
+        if is_valid_domain(dns_name):
+            firewall_result = detect_firewall(dns_name)
+            firewall_results.append(firewall_result)
+        else:
+            firewall_results.append("DNT")
 
     deduplicated_df["firewall"] = firewall_results
 
@@ -249,8 +164,8 @@ if config["run_ip_scan"]:
     deduplicated_df["associated_ips"] = found_ips
 
 if config["http_status_scan"]:
-    print(f"Starting HTTP response scan  . . .")
 
+    print(f"Starting HTTP response scan  . . .")
     http_responses = []  # Store HTTP responses
 
     for dns_name in dns_names_list:
@@ -278,11 +193,8 @@ if config["sec_headers_scan"]:
     # Add a new column "missing_sec_headers" to deduplicated_df
     deduplicated_df["missing_sec_headers"] = sec_header_responses
 
-if config["run_pii_scan"]:
-    print(f"Starting PII scan . . .")
-    scan_and_update_pii_data(deduplicated_df)
-
 if config["run_robots_txt_scan"]:
+
     print(f"Starting robots.txt scan . . .")
     robots_text_response = []
